refactoring remaining tests
