+++ .scalafmt.conf
@@ -1,0 +1,3 @@
+version = "3.7.14"
+maxColumn = 120
+runner.dialect = scala213
